Fix printing of Arch E & D paper sizes [#166192812]
PDF.js was failing when the user printed with the Arch E or Arch D paper size.  This happened because PDF.js created a `canvas` element that exceeded Chrome's maximum area for a `canvas`.

Fixing this by defaulting to 600 DPI but dynamically choosing a lower DPI if 600 would cause PDF.js to exceed Chrome's max area.